ref(auth): Move resolved OIDC provider ownership
Place the resolved provider type and its compile step in provider.rs, where discovery and explicit provider construction already live. This keeps config.rs focused on config shape and mode selection while preserving behavior.
